Fix image loading taking the wrong palette index; do palette lookup in shader
IT WORKS!

Now we just need to provide the actual game palette, and we should be able to
change colours on the fly by changing the palette.

The texture co-ordinates that we use to look up the colour in our palette may
need revisiting. It only works right now because we're clamping the texture
co-ordinates; with the default "repeat" texture wrapping, the background colour
(255) confusingly gets sampled from colour 0.