Remove Image destructor
This was deleting the image data when the image goes out of scope, even though
it is still needed. Our sprite array has a COPY of the object, but `data`
points to the same chunk of memory.

We will need to find a better way to manage this memory - probably using smart
pointers.