Separate Sprite from Unit class
Now a Unit needs only to specify its type, and the UnitRenderer will know how to draw it.

The Rival class owns a list of all Textures. Sprites are created with a reference to their Texture from this list,
and exist in a map, keyed by Unit type. This map is also owned by Rival, but is passed by reference to the UnitRenderer.