Fix transparent backgrounds in portraits
There was a confusing comment in Palette.cpp that claimed that the "Last color is used by entities for transparency",
but this does not seem to be the case. Rather, it is the shader that takes care of entity transparency, by discarding
fragments with a palette index of 1.

This seems fine for now, but if colour 1 is required by any menu textures then we may need to revisit this.