Update docs (+ FreeType logging investigation)
Note that FreeType logging cannot be suppressed without rebuilding the library.
In debug mode, the library is built with the `FT_DEBUG_LEVEL_ERROR` macro defined,
which causes errors to be logged to stderr.

Error occurs here:
https://github.com/freetype/freetype/blob/4d8db130ea4342317581bab65fc96365ce806b77/src/base/ftsystem.c#L261

FT_ERROR is defined here, provided FT_DEBUG_LEVEL_ERROR is defined:
https://github.com/freetype/freetype/blob/4d8db130ea4342317581bab65fc96365ce806b77/include/freetype/internal/ftdebug.h#L290

This sends logs to FT_Message, which prints to stderr:
https://github.com/freetype/freetype/blob/4d8db130ea4342317581bab65fc96365ce806b77/src/base/ftdebug.c#L99

This is not too much of a concern since the release build of the library does not
perform any logging, which is what we link with for our own release build.