Migrate test project to CMake (WIP)
This seems to be *almost* working now (or at least compiling), but a lot of GLEW symbols are missing.
This is probably because we are now including ALL Open-Rival source files indiscriminately, whereas
previously we only included the files that were used by the tests.

We should consider excluding some unneeded files, or stubbing some methods, so that fewer GLEW symbols
are referenced by the tests.