Yet more pathfinding optimisations
1) Divide the pathfinding allowance between all members of a group

This means that the bigger the group, the less thorough their pathfinding will become. They may end up with only
partial paths. Later, we can make units repath when they reach the end of a partial path.

2) Consider the endpoint of a unit's route as a valid goal for subsequent group members

This prevents subsequent units from performing exhaustive pathfinding if the destination is unreachable. This won't
help if units in the group are separated from each other by an untraversable barrier, but this is rare.